// src/enums/CharacterClass.ts
export enum CharacterClass {
  Warrior = 'Warrior',
  Paladin = 'Paladin',
  Hunter = 'Hunter',
  Rogue = 'Rogue',
  Priest = 'Priest',
  Shaman = 'Shaman',
  Mage = 'Mage',
  Warlock = 'Warlock',
  Monk = 'Monk',
  Druid = 'Druid',
  DemonHunter = 'DemonHunter',
  DeathKnight = 'DeathKnight',
  Evoker = 'Evoker',
}

export const CharacterClassArr = Object.values(CharacterClass)